+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 * Creates a PID file containing the Process ID, to be used for package updating.
 * Added an option to "repair" when the `choco.exe.old` problem is encountered.
+* Missing settings are now automatically added to the `settings.json` file with the default value.
 
 ### Changes
 * Minor changes to Advanced menu text and ordering